// front/src/app/swagger/model/loginResponseDto.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginResponseDto { 
    token?: string | null;
    expiresAt?: string;
    email?: string | null;
    role?: string | null;
}