// front/src/app/swagger/model/loginResponseDtoHttpResponse.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HttpStatusCode } from './httpStatusCode';
import { LoginResponseDto } from './loginResponseDto';


export interface LoginResponseDtoHttpResponse { 
    result?: LoginResponseDto;
    readonly message?: string | null;
    statusCode?: HttpStatusCode;
    readonly isError?: boolean;
}
export namespace LoginResponseDtoHttpResponse {
}